refactor: change main entrance to src/index
