+++ test/unit/canvas/core/animate-spec.js
@@ -1,5 +1,5 @@
 const expect = require('chai').expect;
-const G = require('../../../../src/g');
+const G = require('../../../../src/index');
 
 describe('animate', function() {
   const div = document.createElement('div');
