+++ test/unit/canvas/shape/arc-spec.js
@@ -1,5 +1,5 @@
 const expect = require('chai').expect;
-const G = require('../../../../src/g');
+const G = require('../../../../src/index');
 const Util = require('../../../../src/util/index');
 
 const Canvas = G.Canvas;
