+++ test/unit/canvas/shape/cubic-spec.js
@@ -1,6 +1,6 @@
 const expect = require('chai').expect;
 const Canvas = require('../../../../src/canvas');
-const G = require('../../../../src/g');
+const G = require('../../../../src/index');
 const Util = require('../../../../src/util/index');
 const div = document.createElement('div');
 div.id = 'canvas-cubic';
