+++ test/unit/svg/core/animate-spec.js
@@ -1,5 +1,5 @@
 const expect = require('chai').expect;
-const G = require('../../../../src/g');
+const G = require('../../../../src/index');
 
 describe('animate', () => {
   const div = document.createElement('div');
