+++ test/unit/svg/shape/circle-spec.js
@@ -1,5 +1,5 @@
 const expect = require('chai').expect;
-const G = require('../../../../src/g');
+const G = require('../../../../src/index');
 const Canvas = G.Canvas;
 
 const div = document.createElement('div');
