+++ test/unit/svg/shape/clip-spec.js
@@ -1,4 +1,4 @@
-const G = require('../../../../src/g');
+const G = require('../../../../src/index');
 const Canvas = G.Canvas;
 
 const div = document.createElement('div');
